requirements.txt
streamlit>=1.52.2
torch>=2.9.1
torchvision>=0.24.1
pillow
requests
pandas
numpy